+++ src/core/fd_vector.c
@@ -38,6 +38,7 @@
 #include <stdlib.h>
 #include <unistd.h>
 #include <stdio.h>
+#include <stdint.h>
 
 #include "usp_mem.h"
 #include "usp_err.h"